requirements.txt
fastapi==0.115.0
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
mysql-connector-python==8.2.0
python-dotenv==1.0.0